+++ fixtures/base.js
@@ -1,7 +1,6 @@
 import { test as base } from '@playwright/test';
 import LoginPage from '../page_objects/loginPage';
 import SignPage from '../page_objects/signPage';
-import TemplatesActivePage from '../page_objects/templatesActivePage';
 
 const EMAIL = process.env.USER_EMAIL;
 const PASSWORD = process.env.USER_PASSWORD;
@@ -24,13 +23,13 @@ export const test = base.extend({
 
     createNewFolder: [
         async ({ page }, use) => {
-            const templatesActivePage = new TemplatesActivePage(page);
             const signPage = new SignPage(page);
             
-            await signPage.clickTemplateDropdownAndGoTemplatesActivePage();  
-            await templatesActivePage.clickCreateFolderBtn();
-            await templatesActivePage.fillNewFolderNameInputField();
-            await templatesActivePage.clickCreateBtn();
+            const documentsPage = await signPage.clickDocumentsDropdownAndGoDocumentsPage();  
+            await documentsPage.clickCreateFolderBtn();
+            await documentsPage.fillNewFolderNameInputField();
+            await documentsPage.clickCreateBtn();
+            await documentsPage.clickSignaturelyLogoAndGoSignPage();
 
             await use("");
         },
