+++ page_objects/documentsPage.js
@@ -1,5 +1,6 @@
 import SignPage from "./signPage";
 import DocumentsTrashPage from "./documentsTrashPage";
+import { FILL_FOLDER_NAME } from '../testData';
 
 class DocumentsPage {
     constructor(page) {
@@ -17,6 +18,14 @@ class DocumentsPage {
         getYesDeleteBtn: () => this.page.getByRole('button', {name: 'Yes, Delete'}),
         getEmptyTableHeader: () => this.page.locator('.empty-table__header'),
         getTrashSidebarLink: () => this.page.getByRole('link', {name: 'Trash'}),
+        getCreateFolderBtn: () => this.page.getByRole('button', { name: 'Create Folder'}),
+        getNewFolderNameInputField: () => this.page.getByPlaceholder('New Folder Name'),
+        getCreateBtn: () => this.page.getByRole('button', {name: 'Create'}),
+        getOptionsBtn: () => this.page.locator('.documents__optionsDropdown').first(),
+        getToaster: () => this.page.locator('.Toastify__toast-body'),
+        getDeleteBtn: () => this.page.getByRole('button', {name: 'Delete'}),
+        getYesDeteleBtn: () => this.page.getByRole('button', {name: 'Yes, Delete'}),
+        getSignaturelyLogo: () => this.page.locator('a>img'),
     }
 
     async clickSignSidebarLinkAndGoSignPage() {
@@ -55,5 +64,47 @@ class DocumentsPage {
         return new DocumentsTrashPage(this.page);
     }
 
+    async clickCreateFolderBtn() {
+        await this.locators.getCreateFolderBtn().click();
+
+        return this;
+    }
+
+    async fillNewFolderNameInputField() {
+        await this.locators.getNewFolderNameInputField().fill(FILL_FOLDER_NAME) 
+
+        return this;
+    }
+
+    async clickCreateBtn() {
+        await this.locators.getCreateBtn().click();
+
+        return this;
+    }
+
+    async clickOptionsBtn() {
+        await this.locators.getOptionsBtn().click();
+
+        return this;
+    }
+
+    async clickDeleteBtn() {
+        await this.locators.getDeleteBtn().click();
+
+        return this;
+    }
+
+    async clickYesDeleteBtn() {
+        await this.locators.getYesDeteleBtn().click();
+
+        return this;
+    }
+
+    async clickSignaturelyLogoAndGoSignPage() {
+        await this.locators.getSignaturelyLogo().click();
+
+        return new SignPage(this.page);
+    }
 }
+
 export default DocumentsPage;
