+++ tests/06Folders.spec.js
@@ -1,25 +1,24 @@
 import { expect } from "@playwright/test"
 import {test, loginBusinessUser, createNewFolder} from "../fixtures/base.js";
 import SignPage from "../page_objects/signPage";
-import TemplatesActivePage from "../page_objects/templatesActivePage.js";
-const EMAIL = process.env.USER_EMAIL;
-const PASSWORD = process.env.USER_PASSWORD;
-const BASE_URL = process.env.URL;
 
 test.describe('Folders', () => {
 
     test('TC_06_24_01 | Verify the business user can delete folder', async ({ page, loginBusinessUser, createNewFolder }) => {
         const signPage = new SignPage(page); 
-        const templatesActivePage = new TemplatesActivePage(page);
 
-        await templatesActivePage.locators.getToaster().waitFor({ state: 'visible' });
-        await templatesActivePage.locators.getToaster().waitFor({ state: 'hidden' });
+        const documentsPage = await signPage.clickDocumentsDropdownAndGoDocumentsPage();
 
-        await templatesActivePage.clickOptionsBtn();
-        await templatesActivePage.clickDeleteBtn();
-        await templatesActivePage.clickYesDeleteBtn();
-        await templatesActivePage.locators.getToaster().waitFor({ state: 'visible' });
+        await documentsPage.locators.getToaster().waitFor({ state: 'visible' });
+        await documentsPage.locators.getToaster().waitFor({ state: 'hidden' });
 
-        await expect(templatesActivePage.locators.getToaster()).toHaveText('Folder deleted successfully.');
+        await documentsPage.clickOptionsBtn();
+        await documentsPage.clickDeleteBtn();
+        await documentsPage.clickYesDeleteBtn();
+        await documentsPage.locators.getToaster().waitFor({ state: 'visible' });
+
+        await expect(documentsPage.locators.getToaster()).toHaveText('Folder deleted successfully.');
+
+        await documentsPage.clickSignaturelyLogoAndGoSignPage();
     })
 })
